fix!: removing Runtime from global scope
in unity 2021 LTS or later, Runtime is not defined by unity, so a fix was to define it ourselves. but the code was defining it in global scope, so if a 2nd unity webgl instance is created then it would use the Runtime variable from the first one, which causing many errors.

Dropping support for older unity versions means we can go back to using dynCall directly instead of calling via Runtime.dynCall

BREAKING CHANGE: no longer supports unity 2020 or earlier